Update README.md
This is project is w.r.t the Udacity implememntation of EKF (Extended Kalman Filter). In this project I have worked on both LiDAR and RADAR data predict and update the position of a vehicle. As such using KF (Kalman Filter), its not possible to work on non-linear data, hence using EKF the same is linearize.
Following are the "ToDo" task appointed:
1. FusionEKF: It implements EKF for radar and LiDAR sensors
2. kalman_filter: It implements the equations for KF and EKF.'
3. tools: It calculates the RMSE.